services/interaction-service/requirements.txt
Django>=5.2
djangorestframework>=3.15
django-cors-headers>=4.3
psycopg[binary]>=3.2
gunicorn>=22.0

